move `rollback_err` to `parse` module
