+++ leptos-mview-core/src/parse.rs
@@ -75,3 +75,17 @@ pub fn parenthesized<T: Parse>(input: ParseStream) -> syn::Result<(syn::token::P
         Err(input.error("no paren found"))
     }
 }
+
+pub fn rollback_err<F, T>(input: ParseStream, parser: F) -> Option<T>
+where
+    F: Fn(ParseStream) -> syn::Result<T>,
+{
+    let fork = input.fork();
+    match parser(&fork) {
+        Ok(val) => {
+            input.advance_to(&fork);
+            Some(val)
+        }
+        Err(_) => None,
+    }
+}
